ROBOT-1235: Made last changes after general consensus on exact trajectory data structures. Still need to create package for trajectory analyzer core + apps
