+++ mecademic/robot.py
@@ -1,25 +1,46 @@
 #!/usr/bin/env python3
-import logging
+import contextlib
+import copy
+import functools
 import ipaddress
-import time
+import logging
+import queue
+import re
 import socket
 import threading
-import queue
-import functools
-import copy
-import contextlib
-import re
+import time
+from pathlib import PurePath
+
+import pandas as pd
 
 import mecademic.mx_robot_def as mx_def
-from ._robot_common import _MecademicException, _InvalidStateError, _CommunicationError, _DisconnectError
-from ._robot_common import _InterruptException
-from ._robot_common import _InterruptableEvent, _Message, _RobotEvents
-from ._robot_common import _CallbackQueue
-from ._robot_common import _CHECKPOINT_ID_MAX_PRIVATE, _TERMINATE, _string_to_numbers
-from ._robot_data_logger import _RobotDataLogger
+
+from ._robot_trajectory_logger import _RobotTrajectoryLogger
 
 GRIPPER_OPEN = True
 GRIPPER_CLOSE = False
+
+_CHECKPOINT_ID_MAX_PRIVATE = 8191  # Max allowable checkpoint id, inclusive
+
+_TERMINATE = '--terminate--'
+
+
+def _string_to_numbers(input_string):
+    """Convert comma-separated floats in string form to relevant type.
+
+    Parameters
+    ----------
+    input_string : string
+        Comma-separated floats values encoded as a string.
+
+    Returns
+    -------
+    list of numbers
+        Returns converted list of floats or integers, depending on nature of element in 'input_string'.
+
+    """
+
+    return [(float(x) if ('.' in x) else int(x)) for x in input_string.split(',')]
 
 
 def disconnect_on_exception(func):
@@ -39,12 +60,47 @@ def disconnect_on_exception(func):
         except BaseException as e:
             if self._disconnect_on_exception:
                 self.Disconnect()
-                raise _DisconnectError('Automatically disconnected as a result of exception, '
-                                       'set \'disconnect_on_exception\' to False to disable.') from e
+                raise DisconnectError('Automatically disconnected as a result of exception, '
+                                      'set \'disconnect_on_exception\' to False to disable.') from e
             else:
                 raise e
 
     return wrap
+
+
+class MecademicException(Exception):
+    """Base exception class for Mecademic-related exceptions.
+
+    """
+    pass
+
+
+class InvalidStateError(MecademicException):
+    """The internal state of the instance is invalid.
+
+    """
+    pass
+
+
+class CommunicationError(MecademicException):
+    """There is a communication issue with the robot.
+
+    """
+    pass
+
+
+class DisconnectError(MecademicException):
+    """A non-nominal disconnection has occurred.
+
+    """
+    pass
+
+
+class InterruptException(MecademicException):
+    """An event has encountered an error. Perhaps it will never be set.
+
+    """
+    pass
 
 
 class Robot:
@@ -150,7 +206,7 @@ class Robot:
         # Don't allow entering a "with" statement when robot is already connected
         # (since the goal of "with" is to disconnect)
         if self._monitor_handler_thread is not None:
-            raise _InvalidStateError('Robot cannot be connected when entering \'with\' block')
+            raise InvalidStateError('Robot cannot be connected when entering \'with\' block')
         return self
 
     def __exit__(self, exc_type, exc_value, traceback):
@@ -387,14 +443,14 @@ class Robot:
 
         if not (self._monitor_handler_thread and self._monitor_handler_thread.is_alive()):
             self.Disconnect()
-            raise _InvalidStateError('Monitor response handler thread has unexpectedly terminated.')
+            raise InvalidStateError('Monitor response handler thread has unexpectedly terminated.')
 
         if self._offline_mode:  # Do not check rx threads in offline mode.
             return
 
         if not (self._monitor_rx_thread and self._monitor_rx_thread.is_alive()):
             self.Disconnect()
-            raise _InvalidStateError('Monitor rx thread has unexpectedly terminated.')
+            raise InvalidStateError('Monitor rx thread has unexpectedly terminated.')
 
     def _check_command_threads(self):
         """Check that the threads which handle robot command messages are alive.
@@ -405,20 +461,20 @@ class Robot:
 
         if not (self._command_response_handler_thread and self._command_response_handler_thread.is_alive()):
             self.Disconnect()
-            raise _InvalidStateError('No command response handler thread, are you in monitor mode?')
+            raise InvalidStateError('No command response handler thread, are you in monitor mode?')
 
         if self._offline_mode:  # Do not check rx threads in offline mode.
             return
 
         if not (self._command_rx_thread and self._command_rx_thread.is_alive()):
             self.Disconnect()
-            raise _InvalidStateError('No command rx thread, are you in monitor mode?')
+            raise InvalidStateError('No command rx thread, are you in monitor mode?')
 
         # If tx thread is down, attempt to directly send deactivate command to the robot.
         if not (self._command_tx_thread and self._command_tx_thread.is_alive()):
             self._command_socket.sendall(b'DeactivateRobot\0')
             self.Disconnect()
-            raise _InvalidStateError('No command tx thread, are you in monitor mode?')
+            raise InvalidStateError('No command tx thread, are you in monitor mode?')
 
     def _check_internal_states(self):
         """Check that the threads which handle robot messages are alive.
@@ -427,7 +483,7 @@ class Robot:
 
         """
         if self._monitor_mode:
-            raise _InvalidStateError('Cannot send command while in monitoring mode.')
+            raise InvalidStateError('Cannot send command while in monitoring mode.')
         else:
             self._check_command_threads()
 
@@ -491,13 +547,13 @@ class Robot:
             return
 
         if self._command_socket is not None:
-            raise _InvalidStateError('Cannot connect since existing command socket exists.')
+            raise InvalidStateError('Cannot connect since existing command socket exists.')
 
         try:
             self._command_socket = self._connect_socket(self.logger, self._address, mx_def.MX_ROBOT_TCP_PORT_CONTROL)
 
             if self._command_socket is None:
-                raise _CommunicationError('Command socket could not be created. Is the IP address correct?')
+                raise CommunicationError('Command socket could not be created. Is the IP address correct?')
 
             # Create rx thread for command socket communication.
             self._command_rx_thread = self._launch_thread(target=self._handle_socket_rx,
@@ -520,13 +576,13 @@ class Robot:
             return
 
         if self._monitor_socket is not None:
-            raise _InvalidStateError('Cannot connect since existing monitor socket exists.')
+            raise InvalidStateError('Cannot connect since existing monitor socket exists.')
 
         try:
             self._monitor_socket = self._connect_socket(self.logger, self._address, mx_def.MX_ROBOT_TCP_PORT_FEED)
 
             if self._monitor_socket is None:
-                raise _CommunicationError('Monitor socket could not be created. Is the IP address correct?')
+                raise CommunicationError('Monitor socket could not be created. Is the IP address correct?')
 
             # Create rx thread for monitor socket communication.
             self._monitor_rx_thread = self._launch_thread(target=self._handle_socket_rx,
@@ -552,7 +608,7 @@ class Robot:
         except queue.Empty:
             self.logger.error('No response received within timeout interval.')
             self.Disconnect()
-            raise _CommunicationError('No response received within timeout interval.')
+            raise CommunicationError('No response received within timeout interval.')
         except BaseException:
             self.Disconnect()
             raise
@@ -561,7 +617,7 @@ class Robot:
         if response.id != mx_def.MX_ST_CONNECTED:
             self.logger.error('Connection error: {}'.format(response))
             self.Disconnect()
-            raise _CommunicationError('Connection error: {}'.format(response))
+            raise CommunicationError('Connection error: {}'.format(response))
 
         # Attempt to parse robot return data.
         self._robot_info = RobotInfo.from_command_response_string(response.data)
@@ -1114,7 +1170,7 @@ class Robot:
             raise TypeError('Callbacks object is not the appropriate class.')
 
         if self._monitor_handler_thread or self._command_response_handler_thread:
-            raise _InvalidStateError('Callbacks cannot be set if already connected.')
+            raise InvalidStateError('Callbacks cannot be set if already connected.')
 
         self._callback_queue = _CallbackQueue(callbacks)
 
@@ -1138,7 +1194,7 @@ class Robot:
 
         """
         if self._callback_thread:
-            raise _InvalidStateError(
+            raise InvalidStateError(
                 'Cannot call RunCallbacks since callback handler is already running in separate thread.')
 
         # Setting timeout=0 means we don't block on an empty queue.
@@ -2010,7 +2066,7 @@ class Robot:
         with self._main_lock:
             self._check_internal_states()
             if not self._robot_events.on_homed.is_set():
-                raise _InvalidStateError('This command requires robot to be homed.')
+                raise InvalidStateError('This command requires robot to be homed.')
             self._send_command('Delay', [t])
             if self._enable_synchronous_mode:
                 checkpoint = self._set_checkpoint_internal()
@@ -2071,8 +2127,8 @@ class Robot:
         if self._enable_synchronous_mode:
             try:
                 self._robot_events.on_offline_program_started.wait(timeout=timeout)
-            except _InterruptException:
-                raise _InvalidStateError('Offline program start not confirmed. Does program {} exist?'.format(n))
+            except InterruptException:
+                raise InvalidStateError('Offline program start not confirmed. Does program {} exist?'.format(n))
 
     # Non-motion commands.
 
@@ -2113,7 +2169,7 @@ class Robot:
         with self._main_lock:
             if include_timestamp:
                 if not self._robot_info.rt_message_capable:
-                    raise _InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
+                    raise InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
                 else:
                     return copy.deepcopy(self._robot_kinetics.rt_target_joint_pos)
 
@@ -2155,7 +2211,7 @@ class Robot:
         with self._main_lock:
             if include_timestamp:
                 if not self._robot_info.rt_message_capable:
-                    raise _InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
+                    raise InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
                 else:
                     return copy.deepcopy(self._robot_kinetics.rt_target_cart_pos)
 
@@ -2187,7 +2243,7 @@ class Robot:
         with self._main_lock:
             if include_timestamp:
                 if not self._robot_info.rt_message_capable:
-                    raise _InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
+                    raise InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
                 else:
                     return copy.deepcopy(self._robot_kinetics.rt_target_conf)
 
@@ -2219,7 +2275,7 @@ class Robot:
         with self._main_lock:
             if include_timestamp:
                 if not self._robot_info.rt_message_capable:
-                    raise _InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
+                    raise InvalidStateError('Cannot provide timestamp with current robot firmware or model.')
                 else:
                     return copy.deepcopy(self._robot_kinetics.rt_target_conf_turn)
 
@@ -2377,7 +2433,7 @@ class Robot:
 
         """
         if self._file_logger is not None:
-            raise _InvalidStateError('Another file logging operation is in progress.')
+            raise InvalidStateError('Another file logging operation is in progress.')
 
         self.SetMonitoringInterval(monitoringInterval)
         if fields is None:
@@ -2385,19 +2441,19 @@ class Robot:
         else:
             self.SetRealTimeMonitoring(fields)
 
-        self._file_logger = _RobotDataLogger(self._robot_info,
-                                             self._robot_kinetics,
-                                             fields,
-                                             file_path,
-                                             record_time=record_time,
-                                             monitoring_interval=monitoringInterval)
+        self._file_logger = _RobotTrajectoryLogger(self._robot_info,
+                                                   self._robot_kinetics,
+                                                   fields,
+                                                   file_path,
+                                                   record_time=record_time,
+                                                   monitoring_interval=monitoringInterval)
 
     def EndLogging(self):
         """Stop logging robot kinetics to file.
 
         """
         if self._file_logger is None:
-            raise _InvalidStateError('No existing logger to stop.')
+            raise InvalidStateError('No existing logger to stop.')
 
         self._file_logger.end_log()
         self._file_logger = None
@@ -2786,3 +2842,361 @@ class RobotInfo:
                        fw_patch_num=int(matches[5]))
         except Exception as exception:
             raise ValueError('Could not parse robot info string {}'.format(input_string))
+
+
+class _InterruptableEvent:
+    """Extend default event class to also be able to unblock and raise an exception.
+
+    Attributes
+    ----------
+    id : int or None
+        Id for event.
+    _event : event object
+        A standard event-type object.
+    _lock : lock object
+        Used to ensure atomic operations.
+    _interrupted : boolean
+        If true, event is in an error state.
+
+    """
+
+    def __init__(self, id=None, data=None):
+        self._id = id
+        self._data = data
+        self._event = threading.Event()
+        self._lock = threading.Lock()
+        self._interrupted = False
+
+    def wait(self, timeout=None):
+        """Block until event is set or should raise an exception.
+
+        Attributes
+        ----------
+        timeout : float
+            Maximum duration to wait in seconds.
+
+        Return
+        ------
+        success : boolean
+            False if event timed out, true otherwise.
+
+        """
+        success = self._event.wait(timeout=timeout)
+        if self._interrupted:
+            raise InterruptException('Event received exception, possibly because event will never be triggered.')
+        return success
+
+    def wait_for_data(self, timeout=None):
+        """Block until event is set or should raise an exception.
+
+        Attributes
+        ----------
+        timeout : float
+            Maximum duration to wait in seconds.
+
+        Return
+        ------
+        data : object
+            Return the data object.
+
+        """
+        success = self._event.wait(timeout=timeout)
+        if self._interrupted:
+            raise InterruptException('Event received exception, possibly because event will never be triggered.')
+        elif not success:
+            raise InterruptException('Event timed out.')
+        else:
+            return self._data
+
+    def set(self, data=None):
+        """Set the event and unblock all waits. Optionally modify data before setting.
+
+        """
+        with self._lock:
+            self._data = data
+            self._event.set()
+
+    def abort(self):
+        """Unblock any waits and raise an exception.
+
+        """
+        with self._lock:
+            if not self._event.is_set():
+                self._interrupted = True
+                self._event.set()
+
+    def clear(self):
+        """Reset the event to its initial state.
+
+        """
+        with self._lock:
+            self._interrupted = False
+            self._event.clear()
+
+    def is_set(self):
+        """Checks if the event is set.
+
+        Return
+        ------
+        boolean
+            False if event is not set or instance should '_interrupted'. True otherwise.
+
+        """
+        with self._lock:
+            if self._interrupted:
+                return False
+            else:
+                return self._event.is_set()
+
+    def clear_abort(self):
+        """Clears the abort.
+
+        """
+        with self._lock:
+            if self._interrupted:
+                self._interrupted = False
+                self._event.clear()
+
+    @property
+    def id(self):
+        """Make id a read-only property since it should not be changed after instantiation.
+
+        """
+        return self._id
+
+    @property
+    def data(self):
+        """Make data a read-only property and enforce that it is only assignable at construction or using set().
+
+        """
+        return self._data
+
+
+class _Message:
+    """Class for storing a response message from a Mecademic robot.
+
+    Attributes
+    ----------
+    id : integer
+        The id of the message, representing the type of message.
+    data : string
+        The raw payload of the message.
+
+    """
+
+    def __init__(self, id, data):
+        self.id = id
+        self.data = data
+
+    def __repr__(self):
+        return "Message with id={}, data={}".format(self.id, self.data)
+
+    @classmethod
+    def from_string(cls, input):
+        """Construct message object from raw string input.
+
+        Parameters
+        ----------
+        input : string
+            Input string to convert to message.
+
+        """
+        id_start = input.find('[') + 1
+        id_end = input.find(']', id_start)
+        id = int(input[id_start:id_end])
+        # Find next square brackets (contains data).
+        data_start = input.find('[', id_end) + 1
+        data_end = input.find(']', data_start)
+
+        data = ''
+        if data_start != -1 and data_end != -1:
+            data = input[data_start:data_end]
+
+        return cls(id, data)
+
+
+class _RobotEvents:
+    """Class for storing possible status events for the generic Mecademic robot.
+
+    Attributes
+    ----------
+    on_connected : event
+        Set if robot is connected.
+    on_disconnected : event
+        Set if robot is disconnected.
+    on_status_updated : event
+        Set if robot status is updated.
+    on_activated : event
+        Set if robot is activated.
+    on_deactivated : event
+        Set if robot is deactivated.
+    on_homed : event
+        Set if robot is homed.
+    on_error : event
+        Set if robot is in error.
+    on_error_reset : event
+        Set if robot error has been reset.
+    on_p_stop : event
+        Set if robot receives p_stop.
+    on_p_stop_reset : event
+        Set if p_stop is reset.
+    on_motion_paused : event
+        Set if robot motion is paused.
+    on_motion_resumed : event
+        Set if robot motion is not paused.
+    on_motion_cleared : event
+        Set if there are no pending ClearMotion commands.
+    on_activate_sim : event
+        Set if robot is in sim mode.
+    on_deactivate_sim : event
+        Set if robot is not in sim mode.
+    on_conf_updated : event
+        Set if robot configuration has been updated.
+    on_conf_turn_updated : event
+        Set if last joint turn number has been updated.
+    on_joints_updated : event
+        Set if joint angles has been updated.
+    on_pose_updated : event
+        Set if robot pose has been updated.
+    on_brakes_activated : event
+        Set if brakes are activated.
+    on_brakes_deactivated : event
+        Set if brakes are deactivated.
+    on_offline_program_started : event
+        Set if there has been a change in the offline program state.
+    on_end_of_block : event
+        Set if end of block has been reached.
+    on_end_of_cycle: event
+        Set if end of cycle has been reached
+
+    """
+
+    def __init__(self):
+        self.on_connected = _InterruptableEvent()
+        self.on_disconnected = _InterruptableEvent()
+
+        self.on_status_updated = _InterruptableEvent()
+
+        self.on_activated = _InterruptableEvent()
+        self.on_deactivated = _InterruptableEvent()
+
+        self.on_homed = _InterruptableEvent()
+
+        self.on_error = _InterruptableEvent()
+        self.on_error_reset = _InterruptableEvent()
+        self.on_p_stop = _InterruptableEvent()
+        self.on_p_stop_reset = _InterruptableEvent()
+
+        self.on_motion_paused = _InterruptableEvent()
+        self.on_motion_resumed = _InterruptableEvent()
+        self.on_motion_cleared = _InterruptableEvent()
+
+        self.on_activate_sim = _InterruptableEvent()
+        self.on_deactivate_sim = _InterruptableEvent()
+
+        self.on_conf_updated = _InterruptableEvent()
+        self.on_conf_turn_updated = _InterruptableEvent()
+        self.on_joints_updated = _InterruptableEvent()
+        self.on_pose_updated = _InterruptableEvent()
+
+        self.on_brakes_activated = _InterruptableEvent()
+        self.on_brakes_deactivated = _InterruptableEvent()
+
+        self.on_offline_program_started = _InterruptableEvent()
+
+        self.on_end_of_block = _InterruptableEvent()
+        self.on_end_of_cycle = _InterruptableEvent()
+
+        self.on_disconnected.set()
+        self.on_deactivated.set()
+        self.on_error_reset.set()
+        self.on_p_stop_reset.set()
+        self.on_motion_resumed.set()
+        self.on_deactivate_sim.set()
+
+        self.on_status_updated.set()
+        self.on_conf_updated.set()
+        self.on_conf_turn_updated.set()
+        self.on_joints_updated.set()
+        self.on_pose_updated.set()
+        self.on_brakes_activated.set()
+
+    def clear_all(self):
+        """Clear all events.
+
+        """
+        for attr in self.__dict__:
+            self.__dict__[attr].clear()
+
+    def abort_all_except_on_connected(self):
+        """Abort all events, except for on_connected.
+
+        """
+        for attr in self.__dict__:
+            if attr != 'on_connected':
+                self.__dict__[attr].abort()
+
+    def clear_abort_all(self):
+        """Clear aborts for all events.
+
+        """
+        for attr in self.__dict__:
+            self.__dict__[attr].clear_abort()
+
+
+class _CallbackQueue():
+    """Queue class for storing triggered callbacks. Only registered callbacks are added to the queue.
+
+    Attributes
+    ----------
+    _queue : queue
+        Queue to use to store callback names and associated data.
+    _registered_callbacks : set
+        Set of names of registered callbacks.
+
+    """
+
+    def __init__(self, robot_callbacks):
+        self._queue = queue.Queue()
+        self._registered_callbacks = set()
+
+        for attr in robot_callbacks.__dict__:
+            if robot_callbacks.__dict__[attr] is not None:
+                self._registered_callbacks.add(attr)
+
+    def qsize(self):
+        """Returns the queue size.
+
+        """
+        return self._queue.qsize()
+
+    def put(self, callback_name, data=None):
+        """Put the callback name and associated data into the queue if is registered.
+
+        Parameters
+        ----------
+        callback_name : str
+            Name of callback.
+        data : any object type
+            Associated data.
+
+        """
+        if callback_name in self._registered_callbacks or callback_name == _TERMINATE:
+            self._queue.put((callback_name, data))
+
+    def get(self, block=False, timeout=None):
+        """Get the next callback in the queue.
+
+        Parameters
+        ----------
+        block : bool
+            Block on next available callback if true.
+        timeout : float
+            Maximum time to wait on a callback.
+
+        Returns
+        -------
+        tuple of callback name and data
+
+        """
+        return self._queue.get(block=block, timeout=timeout)
